Move PythonCodegen to separate module
